+++ actions/issue-reopener/package.json
@@ -20,7 +20,6 @@
   "dependencies": {
     "@actions/core": "^1.10.0",
     "@actions/exec": "^1.1.1",
-    "@actions/io": "^1.1.3",
     "@actions/tool-cache": "^2.0.1"
   },
   "devDependencies": {
